Correct type of `policy_id` in `incoming_webhook`
According to the upstream documentation [0] and [1] the parameter
for the escalation policy id shall be of type string not integer.
The previous implementation could create resources fine but when
running a plan of an existing resource it would fail with:

```
error: json: cannot unmarshal string into Go struct field incomingWebhook.data.attributes.policy_id of type int
```

This can easily be re-created by running terraform apply to create
a incoming webhook resource with a escalation policy id and then
running terraform plan as this will fail with the error message
mentioned above.

[0] https://betterstack.com/docs/uptime/api/create-incoming-webhook/
[1] https://betterstack.com/docs/uptime/api/incoming-webhooks-response-params/