Add `expiration_policy_id` attribute to `betteruptime_monitor`
The attribute reflects the same attribute in the Monitor API, which was
recently added.

Per discussion with @PetrHeinz, `expiration_policy_id` is not computed
so it is possible to set it to `null` by users. We also always send its
value to the API on create/update, even when it is not set or changed.
This is done because we use Go structs + default JSON serialization,
meaning it's impossible to differentitate between not setting the
attribute and setting it to `null`.